% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_fastqs.R
\name{split_names_fastq2}
\alias{split_names_fastq2}
\title{split_names_fastq2}
\usage{
split_names_fastq2(x,
  format = "{{samplename}}_{{index}}_L00{{lane}}_R{{read}}_{{num}}.fastq.gz")
}
\description{
split_names_fastq2
}

